Some more fixes on small nits and typos (static analysis)

+++ plugins/janus_streaming.c
@@ -4611,7 +4611,7 @@ static void *janus_streaming_relay_thread(void *data) {
 					janus_rtp_header *rtp = (janus_rtp_header *)buffer;
 					ssrc = ntohl(rtp->ssrc);
 					if(source->rtp_collision > 0 && a_last_ssrc && ssrc != a_last_ssrc &&
-							(now-source->last_received_audio) < 1000*source->rtp_collision) {
+							(now-source->last_received_audio) < (gint64)1000*source->rtp_collision) {
 						JANUS_LOG(LOG_WARN, "[%s] RTP collision on audio mountpoint, dropping packet (ssrc=%u)\n", name, ssrc);
 						continue;
 					}
